+++ jmbo_twitter/admin.py
@@ -1,5 +1,5 @@
 from django.contrib import admin
-from django.core.urlresolvers import reverse
+from django.core.urlresolvers import reverse, NoReverseMatch
 
 from jmbo.admin import ModelBaseAdmin
 
@@ -26,13 +26,20 @@ class FeedAdmin(ModelBaseAdmin):
             parent = super(FeedAdmin, self)._actions(obj)
         except AttributeError:
             parent = ''
-        return parent + '''<ul>
-            <li><a href="%s">Fetch tweets</a></li>
-            <li><a href="%s">View tweets</a></li>
-            </ul>''' % (
-                reverse('feed-fetch-force', args=[obj.name]), 
-                reverse('feed-tweets', args=[obj.name])
-            )
+
+        try:
+            url = reverse('feed-fetch-force', args=[obj.name]), 
+            one = '<a href="%s">Fetch tweets</a>' % url
+        except NoReverseMatch:
+            one = "Fetch tweets - add jmbo_twitter admin_urls to settings, eg. <code>(r'^admin/', include('jmbo_twitter.admin_urls'))</code>"
+
+        try:
+            url = reverse('feed-tweets', args=[obj.name]), 
+            two = '<a href="%s">View tweets</a>' % url
+        except NoReverseMatch:
+            two = "View tweets - add jmbo_twitter admin_urls to settings, eg. <code>(r'^admin/', include('jmbo_twitter.admin_urls'))</code>"
+
+        return parent + '<ul><li>' + one + '</li><li>' + two + '</li></ul>'
     _actions.short_description = 'Actions'
     _actions.allow_tags = True