+++ jmbo_twitter/admin_urls.py
@@ -5,14 +5,14 @@ urlpatterns = patterns(
     '',
 
     (
-        r'^admin/jmbo_twitter/(?P<name>[\w-]+)/fetch-force/$',  
+        r'jmbo_twitter/feed/(?P<name>[\w-]+)/fetch-force/$',  
         'jmbo_twitter.admin_views.feed_fetch_force',
         {'redirect_to' : '/admin/jmbo_twitter/feed'},
         'feed-fetch-force',
     ),
 
     (
-        r'^admin/jmbo_twitter/(?P<name>[\w-]+)/tweets/$',  
+        r'^jmbo_twitter/feed/(?P<name>[\w-]+)/tweets/$',  
         'jmbo_twitter.admin_views.feed_tweets',
         {},
         'feed-tweets',