+++ jmbo_twitter/admin_views.py
@@ -2,6 +2,8 @@ from django.contrib.admin.views.decorators import staff_member_required
 from django.http import HttpResponseRedirect
 from django.shortcuts import get_object_or_404, render_to_response
 from django.template import RequestContext
+from django.contrib import messages
+from django.utils.translation import ugettext as _
 
 from jmbo_twitter.models import Feed
 
@@ -11,9 +13,8 @@ def feed_fetch_force(request, name, redirect_to):
     """Forcibly fetch tweets for the feed"""
     feed = Feed.objects.get(name=name)
     feed.fetch(force=True)
-    request.user.message_set.create(
-        message="Fetched tweets for %s" % feed.name
-    )
+    msg = _("Fetched tweets for %s" % feed.name)
+    messages.success(request, msg, fail_silently=True)
     return HttpResponseRedirect(redirect_to)